[HUDI-3567] Refactor HoodieCommonUtils to make code more reasonable (#4982)
